==> workflow/03-estimate_lux.Rmd <==
---
title: "03-estimate_lux"
date: "Compiled at `r format(Sys.time(), '%Y-%m-%d %H:%M:%S', tz = 'UTC')` UTC"
output: github_document
params:
  name: "03-estimate_lux" # change if you rename file
---

```{r here, message=FALSE}
here::i_am(paste0(params$name, ".Rmd"), uuid = "4eb39ded-b9ea-43e0-b6e3-fcdc2284fdf4")
```

The purpose of this document is ...

```{r packages}
library(tidyverse)
library(pracma)
library(foreach)
```

```{r directories}
# create or *empty* the target directory, used to write this file's data: 
projthis::proj_create_dir_target(params$name, clean = TRUE)

# function to get path to target directory: path_target("sample.csv")
path_target <- projthis::proj_path_target(params$name)

# function to get path to previous data: path_source("00-import", "sample.csv")
path_source <- projthis::proj_path_source(params$name)
```

## Read previous data. 

 
```{r}
filter.metadata <- read_tsv(path_source("02-growth_filtering", "filter_metadata.tsv.gz"), show_col_types = FALSE)


od_10h_data <- read_tsv(path_source("02-growth_filtering", "od_10hours.tsv.gz"), show_col_types = FALSE) %>% 
  rename("experiment_id" = "curve_id") %>% 
  select(experiment_id, od_10h)
```


## Read Lux data

```{r}
read_lux_files <- function(file_name, input_directory){
  
  # Read file
  file_path <- file.path(input_directory, file_name)
  timeseries_df <- read.table(file_path, sep='\t', header = TRUE) 
  
  # Add timepoint column
  timeseries_df <- timeseries_df %>%
    mutate(timepoint = 1:n())
  
  # Add time spent in incubator (in hours)
  ## Add time 0
  t0 <- strptime(timeseries_df$Time[1],  format = "%m/%d/%Y %H:%M:%S", tz = "CET")
  
  timeseries_df <- timeseries_df %>% 
    ## Format time-stamps
    mutate(time_format = strptime(Time,  format = "%m/%d/%Y %H:%M:%S", tz = "CET"),
           
           # Time spent in incubator
           time_spent_h = as.numeric(time_format - t0, units="hours")) %>% 
    
    select(-time_format)
  
  # Long format
  timeseries.long <- timeseries_df %>% 
    pivot_longer(cols = -c(timepoint, Time, time_spent_h), names_to = "well", values_to = "lux")
  
  
  # Add information
  assay.info <- str_split(file_name, "/")
  batch_name <- assay.info[[1]][1]
  
  plate_id <- assay.info[[1]][2]
  plate_id <- str_remove(plate_id, "_LUX.tsv.gz")
  
  
  timeseries.long <- timeseries.long %>% 
    mutate(plate_id = plate_id) %>% 
    
    separate(plate_id, sep = "_", into = c("promoter", "lib_plate", "replicate")) %>% 
  
  return(timeseries.long)
}
```


```{r}
lux_files <- list.files(path_source("00-import/Salmonella/lux_data"), recursive = TRUE)
# Do not consider validation screen files yet. 
lux_files <- lux_files[!str_starts(lux_files, "validataion_screen/")] 

# Read files sequentially
start <- proc.time()[3]
registerDoSEQ()
complete_lux <- foreach(fname = lux_files, .combine = "rbind") %dopar% read_lux_files(fname, path_source("00-import/Salmonella/lux_data"))
end <- proc.time()[3]

end - start / 60
```


Take from third timepoint to be consistent with growth analysis

```{r}
complete_lux <- complete_lux %>% 
  filter(timepoint >= 3)
```


```{r}
complete_lux <- complete_lux %>% 
  unite("experiment_id", c(promoter, lib_plate, well, replicate))
```



## Estimate AUC. 
  
```{r}
# Closest timepoint to 10h
complete_lux.10tp <- complete_lux %>% 
  mutate(diff10 = abs(time_spent_h - 10)) %>% 
  group_by(experiment_id) %>% 
  filter(diff10 == min(diff10)) %>% 
  ungroup() %>% 
  select(experiment_id, time_spent_h) %>% 
  rename("tp_10h" = "time_spent_h")
```


```{r}
lux.auc <- complete_lux %>% 
  left_join(complete_lux.10tp, by="experiment_id") %>% 
  filter(time_spent_h <= tp_10h) %>% 
  
  group_by(experiment_id) %>% 
  summarise(lux.auc.10h = trapz(x=time_spent_h, y=lux),
            lux.auc.10h.tp = trapz(x=timepoint, y=lux)) %>% 
  ungroup()
```


## Normalized Lux AUC. 

```{r}
lux.auc.complete <- lux.auc %>% 
  left_join(od_10h_data, by="experiment_id") %>% 
  mutate(lux.normed = lux.auc.10h / od_10h,
         log2.lux.normed = log2(lux.normed))


```

We can then continue with the filtered experiments. We can add an additional filter where we know that for feoB, lp1 and lp2 failed. 
  
```{r}
filter.metadata <- filter.metadata %>% 
  unite("promoter_libplate", c(promoter, libplate), remove=FALSE) %>% 
  
  mutate(filter.status = case_when(filter.status != "remove" & promoter_libplate %in% c("PfeoB_lp1", "PfeoB_lp2") ~ "remove",
                                   TRUE ~ filter.status),
         
         filter.reason = case_when(filter.reason == "None" & promoter_libplate %in% c("PfeoB_lp1", "PfeoB_lp2") ~ "failed.reporter",
                                   TRUE ~ filter.reason))

filter.metadata %>% count(filter.reason)
```




```{r}
kept.experiments <- filter.metadata %>% filter(filter.status == "keep") %>% pull(curve_id)

lux.auc.filt <- lux.auc.complete %>% 
  filter(experiment_id %in% kept.experiments) %>% 
  
  separate(experiment_id, c("promoter", "libplate", "well", "replicate"), remove=FALSE) %>% 
  unite("srn_code", c(libplate, well))
```


```{r}
normed.lux.p <- lux.auc.filt %>% 
  unite("promoter_replicate", c(promoter, replicate), remove=FALSE) %>% 
  
  ggplot(aes(y=log2.lux.normed, x=promoter_replicate, color=promoter)) +
  geom_boxplot() +
  
  theme_bw() +
  labs(x="Promoter_Replicate",
       y=latex2exp::TeX("$log_2$ (Normalized Lux)")) +
  
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        legend.position = "none")

normed.lux.p
ggsave(path_target("log2_normalized_lux_bp.png"), plot= normed.lux.p, dpi = 300, width = 10)



normed.lux.p <- lux.auc.filt %>% 
  unite("promoter_replicate", c(promoter, replicate), remove=FALSE) %>% 
  
  ggplot(aes(y=lux.normed, x=promoter_replicate, color=promoter)) +
  geom_boxplot() +
  
  theme_bw() +
  labs(x="Promoter_Replicate",
       y=latex2exp::TeX("Normalized Lux")) +
  
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        legend.position = "none")

ggsave(path_target("normalized_lux_bp.png"), plot= normed.lux.p, dpi = 300, width = 10)

```

## Median centering. 
  
To bring all promoters to a similar scale. 
  
```{r}
lux_medians <- lux.auc.filt %>% 
  unite("promoter_replicate", c(promoter, replicate), sep="_") %>% 
  
  group_by(promoter_replicate) %>% 
  summarise(median.log2.lux.normed = median(log2.lux.normed, na.rm = TRUE))


lux.auc.filt <- lux.auc.filt %>% 
  unite("promoter_replicate", c(promoter, replicate), sep="_", remove=FALSE) %>% 
  
  left_join(lux_medians, by="promoter_replicate") %>% 
  
  mutate(log2.lux.normed.centered = log2.lux.normed - median.log2.lux.normed) %>% 
  
  select(-promoter_replicate)



normed.lux.p <- lux.auc.filt %>% 
  unite("promoter_replicate", c(promoter, replicate), remove=FALSE) %>% 
  
  ggplot(aes(y=log2.lux.normed.centered, x=promoter_replicate, color=promoter)) +
  geom_boxplot() +
  
  theme_bw() +
  labs(x="Promoter_Replicate",
       y=latex2exp::TeX("Centered $log_2$ (Normalized Lux)")) +
  
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        legend.position = "none")

normed.lux.p
ggsave(path_target("centered_log2_normalized_lux_bp.png"), plot= normed.lux.p, dpi = 300, width = 10)
```

## Comparing replicate values

```{r}
comparison.normed.lux.p <- lux.auc.filt %>% 
  select(promoter, srn_code, replicate, log2.lux.normed.centered) %>% 
  
  pivot_wider(id_cols = c(promoter, srn_code), names_from = replicate, values_from = log2.lux.normed.centered) %>% 
  
  
  
  ggplot(aes(x=r1, y=r2)) +
  geom_point(alpha=0.5, size=1) +
  geom_abline(color="orange", linetype="longdash") +
  geom_smooth(formula = y~x, method="lm", color="red") +
  facet_wrap(~promoter) +
  theme_bw() +
  
  labs(x="Replicate 1",
       y="Replicate 2",
       title = "Comparison of centered normalized Lux")

comparison.normed.lux.p

ggsave(path_target("comparison_normed_lux.png"), dpi=300, plot=comparison.normed.lux.p)
```



## Write files. 
  
```{r}
write_tsv(filter.metadata, path_target("filter_metadata.tsv.gz"))
write_tsv(lux.auc.complete, path_target("lux_auc_complete.tsv.gz"))
write_tsv(lux.auc.filt, path_target("lux_auc_filtered_median.tsv.gz"))
```


## Files written

These files have been written to the target directory, ```r paste0("data/", params$name)```:

```{r list-files-target}
projthis::proj_dir_info(path_target())
```
